Split up code into separate packages
