+++ package.json
@@ -1,10 +1,7 @@
 {
-  "name": "@openapi-ts/backend",
+  "name": "openapi-ts-backend",
   "version": "2.0.5",
   "description": "",
-  "bin": {
-    "openapi-ts-backend": "dist/tools/cli.js"
-  },
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "homepage": "https://github.com/henhal/openapi-ts-backend#readme",
@@ -13,21 +10,15 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/henhal/openapi-ts-backend"
+    "url": "git+https://github.com/henhal/openapi-ts-backend.git"
   },
   "scripts": {
-    "build": "tsc",
-    "clean": "rm -rf dist && rm -f tsconfig.tsbuildinfo",
-    "lint": "eslint src --ext js,ts --max-warnings 0",
-    "postbuild": "chmod u+x dist/tools/cli.js",
-    "prepublish": "yarn clean && yarn build && yarn test",
-    "prebuild": "rm -rf src/test/gen",
-    "pretest": "MODULE_PATH=../.. ts-node src/tools/cli.ts generate-types src/test/api.yml src/test/gen",
-    "test": "NODE_PATH=src LOG_LEVEL=${LOG_LEVEL:=error} jest --config src/test/jest.config.js"
+    "lint": "eslint src --ext js,ts --max-warnings 0"
   },
   "author": "henrik.hall2@gmail.com",
   "files": [
-    "dist/"
+    "dist/",
+    "README.md"
   ],
   "keywords": [
     "openapi",
@@ -46,32 +37,11 @@
     "lambda"
   ],
   "dependencies": {
-    "@openapi-ts/request-types": "^1.0.5",
-    "ajv": "^8.17.1",
-    "ajv-formats": "^3.0.1",
-    "js-yaml": "^4.0.0",
-    "json-schema-to-ts": "^3.1.1",
-    "loglevel": "^1.7.1",
-    "openapi-backend": "^5.15.0",
-    "openapi-types": "^12.1.3",
-    "openapi-typescript": "^7.9.1"
-  },
-  "devDependencies": {
-    "@types/jest": "^30.0.0",
-    "@types/js-yaml": "^4.0.0",
-    "@types/node": "^24.7.0",
-    "@typescript-eslint/eslint-plugin": "^8.45.0",
-    "@typescript-eslint/parser": "^8.45.0",
-    "eslint": "^9.37.0",
-    "jest": "^30.2.0",
-    "ts-jest": "^29.4.4",
-    "ts-node": "^10.9.2",
     "typescript": "^5.9.3"
   },
-  "peerDependencies": {
-    "typescript": ">=3"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
+  "workspaces": [
+    "packages/lib",
+    "packages/cli",
+    "packages/test"
+  ]
 }
