+++ packages/cli/src/typegen/templates.ts
