+++ packages/lib/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "@openapi-ts/backend",
+  "version": "3.0.1",
+  "scripts": {
+    "build": "tsc"
+  },
+  "main": "dist/index.js",
+  "files": [
+    "dist"
+  ],
+  "types": "dist/index.d.ts",
+  "dependencies": {
+    "ajv": "^8.17.1",
+    "ajv-formats": "^3.0.1",
+    "js-yaml": "^4.0.0",
+    "json-schema-to-ts": "^3.1.1",
+    "loglevel": "^1.7.1",
+    "openapi-backend": "^5.15.0",
+    "openapi-types": "^12.1.3",
+    "openapi-typescript": "^7.9.1"
+  },
+  "devDependencies": {
+    "@openapi-ts/request-types": "^1.0.5",
+    "@types/js-yaml": "^4.0.0",
+    "@types/node": "^24.7.0",
+    "@typescript-eslint/eslint-plugin": "^8.45.0",
+    "@typescript-eslint/parser": "^8.45.0",
+    "eslint": "^9.37.0",
+    "ts-node": "^10.9.2"
+  },
+  "peerDependencies": {
+    "typescript": "^5.9.3"
+  }
+}
