+++ packages/lib/src/errors.ts
