+++ packages/lib/src/openapi.ts
