+++ packages/test/jest.config.ts
@@ -1,0 +1,16 @@
+/*
+ * For a detailed explanation regarding each configuration property and type check, visit:
+ * https://jestjs.io/docs/configuration
+ */
+
+export default {
+    transform: {
+        '^.+\\.ts?$': 'ts-jest',
+    },
+    clearMocks: true,
+    collectCoverage: true,
+    coverageDirectory: 'coverage',
+    coverageProvider: 'v8',
+    testMatch: ['**.test.ts'],
+    // setupFiles: ['./tests/setEnvVars.js'],
+};
