+++ packages/test/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "noEmit": true,
+    "target": "ESNext",
+    "lib": ["ES2022"],
+  },
+}
