--- src/test/jest.config.js
@@ -1,13 +1,0 @@
-module.exports = {
-  "setupFilesAfterEnv": [
-    "./jest.setup.js"
-  ],
-  "testMatch": [
-    "**/__tests__/**/*.+(ts|tsx|js)",
-    "**/?(*.)+(test).+(ts|tsx|js)"
-
-  ],
-  "transform": {
-    "^.+\\.(ts|tsx)$": "ts-jest"
-  },
-};
