--- src/test/jest.setup.js
@@ -1,1 +1,0 @@
-jest.setTimeout(20000);
